pyproject.toml
[tool.poetry]
name = "fasthtml-recipes"
version = "0.1.0"
description = "FastHTML recipes"
authors = ["Philip Nuzhnyi <philip.nuzhnyy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-fasthtml = "^0.6.12"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
